Move internal source to library
This because is more simple link all code in every test binary.
