+++ src/internal/ConverterGuiProxy.cpp
