+++ src/internal/TranslationObject.hpp
