+++ tests/tst_csv2ts.cpp
@@ -1,4 +1,4 @@
-#include "../src/ConverterFactory.hpp"
+#include "ConverterFactory.hpp"
 
 #include <QFile>
